refactor: remove unnecessary abstract class and refactor
